Organize useFetchSingleDataByKey to make it extendable
